requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
fuzzywuzzy
python-Levenshtein
ultralytics
pillow
pandas
git+https://github.com/openai/CLIP.git
tqdm
fastapi
uvicorn
python-multipart